create suggestion component
